pyproject.toml
[project]
name = "data-analytics-pipeline"
version = "1.0.0"
description = "PostgreSQL Data Engineering Pipeline for SmartMoving CRM Data Consolidation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "splink[postgres]>=3.0.0",
    "great-expectations>=0.18.0",
    "prisma>=0.11.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "gender-guesser>=0.4.0",
    "python-Levenshtein>=0.21.0",
    "pyarrow>=12.0.0",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

